\docType{class}
\name{Animanipulate-class}
\alias{Animanipulate-class}
\title{Reference class for manipulate object
Expands functionality of the Manipulate object from John Verzani's gWidgetsManipulate package}
\description{
Reference class for manipulate object Expands functionality
of the Manipulate object from John Verzani's
gWidgetsManipulate package
}
\section{Methods}{
\itemize{
\item \code{change_handler(...)}: Evaluate code with current values

\item \code{execute(w = gwindow(gettext("ManipulateR"), visible = FALSE))}: Make the GUI

\item \code{execute()}: Make the GUI

\item \code{get_values()}: Get widget values as list

\item \code{validate_controls()}: Validate that controls are specified properly
}}

