\name{animate}
\alias{animate}
\title{Command to go inside a knitr code chunk (or inside an actual animate call)
--> made by autogeneration, ... have to be just regular character arrays describing the function variables (but can vary in multiple dimension!)
--> all parameters for the animation should go into the surrounding knit_r code block}
\usage{
animate(_expr, ...)
}
\description{
Command to go inside a knitr code chunk (or inside an
actual animate call) --> made by autogeneration, ... have
to be just regular character arrays describing the function
variables (but can vary in multiple dimension!) --> all
parameters for the animation should go into the surrounding
knit_r code block
}

