\name{checkbox}
\alias{checkbox}
\title{Checkbox control constructor}
\usage{
checkbox(initial = FALSE, label = "")
}
\arguments{
  \item{initial}{logical checked or unchecked}

  \item{label}{character If non-"", labels checkbox}
}
\value{
Checkbox instance
}
\description{
Checkbox control constructor
}

