\name{mapVectorToControl.numeric}
\alias{mapVectorToControl.numeric}
\title{numeric to  slider}
\usage{
\method{mapVectorToControl}{numeric}(x)
}
\description{
numeric to slider
}
\note{
could expand by mapping single numeric values to text input
with coerce=as.numeric
}
\author{
John Verzani (gWidgetsManipulate package)
}

