\name{mapVectorToControl}
\alias{mapVectorToControl}
\title{Map a vector to a control}
\usage{
mapVectorToControl(x)
}
\arguments{
  \item{x}{vector defining control}
}
\value{
an instance of ManipulateControls or an error
}
\description{
Map a vector to a control
}
\author{
John Verzani (gWidgetsManipulate package)
}

