\name{picker}
\alias{picker}
\title{Picker  widget}
\usage{
picker(..., initial = NULL, label = "", multiple = FALSE)
}
\arguments{
  \item{...}{list of values separated by comma}

  \item{initial}{initial chocie}

  \item{label}{label for widget}

  \item{multiple}{logical. If \code{TRUE} use combobox,
  \code{FALSE} use table with checkbox}
}
\value{
\code{Picker} object
}
\description{
Widget to select a single character value from list or
select values from list
}

